Accordion refactor
