+++ gatsby-cmketech/src/assets/sass/components/_accordions.scss
@@ -1,6 +1,13 @@
-.accordion {
+.accordion, .accordion--active {
     width: 85%;
     margin: 2rem auto;
+
+    &--active {
+        width: 85%;
+        margin: 2rem auto;
+        border: 1px solid $medium_gray;
+        padding: 1rem;
+    }
 
     h5 {
         @media #{$mq-largeScreen} {
@@ -21,16 +28,29 @@
         @media #{$mq-largeScreen} {
             display: flex;
             align-items: center;
-            padding: 0.5rem 1rem;
             justify-content: space-between;
+            padding: 0.5rem 1rem;
+            margin: 1rem 0;
             color: $white;
             background-color: $navy-blue;
         }
     }
 
+    .accordion--header__active {
+        @media #{$mq-largeScreen} {
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
+            background-color: $white;
+        }
+
+        h5 {
+            color: $navy_blue;
+        }
+    }
+
     .accordion--content {
         @media #{$mq-largeScreen} {
-            border: 1px solid $medium_gray;
             margin-top: 1.5rem;
             padding: 0.75rem 1rem;
         }
