+++ gatsby-cmketech/src/assets/sass/components/_image-grid.scss
@@ -47,7 +47,8 @@
             padding: 0 7rem;
             width: 75%;
             margin: 0 auto;
-            grid-template-columns: repeat(6, 1fr);
+            grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+            // grid-template-columns: repeat(6, 1fr);
             grid-template-rows: 20%;
             grid-gap: 2rem;
             grid-template-areas:
